Get auth-block specific verifications + storages working
Using special verification or storages by auth-block was not
tested and had a couple of implementation errors.

This commits adds basic tests for this feature, fixes the
issues and does some smaller features adjustments (e.g. if
there is only one storage or verification these are used as
default one).

Fix typos in dns01-dnsUpdate implementation.