Run boulder via docker-compose
Boulder is a complexe software with multiple components. Lets use their
maintained docker-compose setup to get a running development version to
test our software against it.

We want to focus on acme-mgmtserver not how to setup boulder.

With this change, boulder connects to the http listener from docker,
so listening only on 127.0.0.1 does not work anymore. Listen on
 FAKE_DNS ip - that is the IP that boulder uses for connects.