Test with different ACME backends
`pebble` is a dedicated (small) project to provided an ACME server for
testing clients: it is often stricter as it must not support real-life
usecases.

Execute with `boulder` and `pebble`.

The challtest server supports to define the default address used to
check HTTP challenges via a HTTP call.

Lets define this value during test bootstrap; this reduces on adjustment
needed during docker-compose startup.