Simplify crypto dependencies
Only use cryptography and OpenSSL (better exceptions) to parse CSR.

`ndg.httpsclient` should be needed in Python 2.7. `pyasn` might be a
transitive dependecy but it is nothing we care about.

`cryptography` has important helper methods to extract the subject
alternative name extensions. The CSR parsing in `cryptography` might
raise generic `ValueError`, so we stick with PyOpenSSL for parsing for
now. The converation method `to_cryptography` has only added in
PyOpenSSL 17.1.0 - so this gets the new minimal required version.