pebble: disable badNonces
Pebble response with badNonce randomly (based on a percentage). So even
if the client does a retry, it can still get again a badNonce. This is
unexcepted (after the first try the nonce should be valid). This is not
supported by acme (it only does on retry). Disable it completely for
now.